third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-trimmed-attribute-value.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.selectNodeWithId('inspected', execute);

  function execute() {
    var treeElement = ElementsTestRunner.firstElementsTreeOutline().findTreeElement(
        ElementsTestRunner.expandedNodeWithId('inspected'));
    var textElement = treeElement.listItemElement.getElementsByClassName('webkit-html-attribute')[0];
    TestRunner.addResult('Original textContent');
    TestRunner.addResult(treeElement.title.textContent);

    treeElement._startEditingTarget(textElement);
    TestRunner.addResult('textContent when editing \'href\'');
    TestRunner.addResult(treeElement.title.textContent);

    eventSender.keyDown('Tab');
    TestRunner.addResult('textContent after moving to \'id\'');
    TestRunner.addResult(treeElement.title.textContent);

    textElement = treeElement.listItemElement.getElementsByClassName('webkit-html-attribute')[1];
    textElement.dispatchEvent(TestRunner.createKeyEvent('Escape'));
    TestRunner.addResult('textContent after canceling the edit (equal to the original one)');
    TestRunner.addResult(treeElement.title.textContent);

    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that user can mutate DOM by means of elements panel.
</p>

<div>
    <a href="data:text/plain;,12345678901234567890123456789012345678901234567890123456789012345678901234567890/123456789012345678901234567890123456789012345678901234567890" id="inspected">Anchor</a>
</div>
</body>
</html>
