third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-trimmed-attribute-value-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that user can mutate DOM by means of elements panel.

Original textContent
<a href=​"data:​text/​plain;​,1234567890…23456789012345678901234567890" id=​"inspected">​Anchor​</a>​
textContent when editing 'href'
<a href="data:text/plain;,12345678901234567890123456789012345678901234567890123456789012345678901234567890/123456789012345678901234567890123456789012345678901234567890" id=​"inspected">​Anchor​</a>​
textContent after moving to 'id'
<a href=​"data:​text/​plain;​,1234567890…23456789012345678901234567890" id="inspected">​Anchor​</a>​
textContent after canceling the edit (equal to the original one)
<a href=​"data:​text/​plain;​,1234567890…23456789012345678901234567890" id=​"inspected">​Anchor​</a>​
